refactor: move Lift to module and remove cycle_time
